change Server and Database Setting Tool from go-ini to go-viper
