+++ yaml-validation/conf/config.yml
@@ -1,0 +1,14 @@
+Server:
+  HttpPort: :8180
+
+Database:
+  Type: mongodb
+  User: beeyond
+  Password: beeyond
+  Host: localhost
+  Port: 27017
+  Name: beeyond_validation_db
+
+KubernetesJsonschema:
+  KubernetesVersion: v1.17.0
+  Url: https://kubernetesjsonschema.dev
